% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{augment.eglm}
\alias{augment.eglm}
\title{Augment a(n) eglm object}
\usage{
augment.eglm(
  x,
  data = model.frame(x),
  newdata = NULL,
  type.predict = c("link", "response", "terms"),
  type.residuals = c("deviance", "pearson"),
  se_fit = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{eglm} object.}

\item{data}{A data.frame/tibble to add observation-level information to,
returning a data.frame object with the same number of rows as data.}

\item{newdata}{Following the same conventions as the data argument, except
with the underlying assumption that the model has not “seen” the data yet.}

\item{type.predict}{offers choice between link, response and terms}

\item{type.residuals}{Offers choice between deviance and pearson}

\item{se_fit}{Boolean logic to add standard error of fitted values}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A data.frame giving summary statistics on the residuals
}
\description{
Augment a(n) eglm object
}
\examples{
# load the mtcars dataset
data(mtcars)

# fit a generalised linear model
mtcars_model <- eflm::eglm(am ~ wt, mtcars, family = "binomial")

# data.frame returning summary statistics on the residuals
augment(mtcars_model)

}
