% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{augment.elm}
\alias{augment.elm}
\title{Augment a(n) elm object}
\usage{
augment.elm(
  x,
  data = model.frame(x),
  newdata = NULL,
  se_fit = FALSE,
  interval = c("none", "confidence", "prediction"),
  ...
)
}
\arguments{
\item{x}{A \code{elm} object.}

\item{data}{A data.frame/tibble to add observation-level information to,
returning a data.frame object with the same number of rows as data.}

\item{newdata}{Following the same conventions as the data argument, except
with the underlying assumption that the model has not “seen” the data yet.}

\item{se_fit}{Boolean logic to add standard error of fitted values}

\item{interval}{Interval options in a linear model}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A data.frame giving summary statistics on the residuals
}
\description{
Augment a(n) elm object
}
\examples{
# load the trees dataset
data(trees)

# fit a linear model on timber volume
trees_model <- eflm::elm(Volume ~ Girth + Height, data = trees)

# data.frame returning summary statistics on the residuals
augment(trees_model)

}
