% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{deviance.eglm}
\alias{deviance.eglm}
\title{deviance method for eglm fitted objects}
\usage{
\method{deviance}{eglm}(object, ...)
}
\arguments{
\item{object}{eglm fitted object}

\item{...}{not used}
}
\value{
The value of the deviance extracted from the object
}
\description{
deviance method for eglm fitted objects
}
