% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflm-package.R
\docType{package}
\name{eflm-package}
\alias{eflm}
\alias{eflm-package}
\title{eflm: Efficient Fitting of Linear and Generalized Linear Models}
\description{
Efficient Fitting of Linear and Generalized Linear Models by using
  just base R. As an alternative to lm() and glm(), this package provides elm()
  and eglm(), with a significant speedup when the number of 
  observations is larger than the number of parameters to estimate. The speed
 gains are obtained by reducing the NxP model matrix to a PxP matrix, and the 
 best computational performance is obtained when R is linked against OpenBLAS,
 Intel MKL or other optimized BLAS library. This implementation aims at being
 compatible with 'broom' and 'sandwich' packages for summary statistics and
 clustering by providing S3 methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachamaltese/eflm/}
  \item Report bugs at \url{https://github.com/pachamaltese/eflm/issues/}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Alexey Kravchenko [thesis advisor]
  \item Constanza Prado [thesis advisor]
  \item Yoto Yotov [thesis advisor]
  \item Neal Fulz (independent code auditing) [reviewer]
  \item Rahul Saxena (improved broom integration) [contributor]
  \item  The United Nations [data contributor, copyright holder, funder]
}

}
\keyword{internal}
