% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eglm-control.R
\name{eglm.control}
\alias{eglm.control}
\title{Auxiliary for Controlling GLM Fitting}
\usage{
eglm.control(epsilon = 1e-08, maxit = 25, trace = FALSE)
}
\arguments{
\item{epsilon}{positive convergence tolerance \eqn{\epsilon}; the iterations
converge when \eqn{|dev - dev_{old}|/(|dev| + 0.1) < \epsilon}.}

\item{maxit}{integer giving the maximal number of IWLS iterations.}

\item{trace}{logical indicating if output should be produced for each
iteration.}
}
\description{
Auxiliary function for \code{eglm} fitting. Typically only used internally by
\code{glm.wfit}, but may be used to construct a control argument to either
function.
}
