% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{eglm}
\alias{eglm}
\title{fast generalized linear model fitting}
\usage{
eglm(model, ...)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of model specification are given under \sQuote{Details}.}

\item{family}{a description of the error distribution and link function to be
used in the model. This can be a character string naming a
family function, a family function or the result of a call to a family
function. See \code{\link{family}} for details of family functions.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link{as.data.frame}} to a data frame) containing the variables in
the model. If not found in \code{data}, the variables are taken from
\code{environment(formula)}, typically the environment from which \code{lm}
is called.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be \code{NULL} or a numeric vector. If non-NULL, weighted
least squares is used with weights \code{weights} (that is, minimizing
\code{sum(w*e^2)}); otherwise ordinary least squares is used.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{etastart}{starting values for the linear predictor.}

\item{mustart}{starting values for the vector of means.}

\item{offset}{this can be used to specify an \emph{a priori} known component
to be included in the linear predictor during fitting. This should be
\code{NULL} or a numeric vector or matrix of extents matching those of the
response. One or more \code{\link{offset}} terms can be included in the
formula instead or as well, and if more than one are specified their sum is
used. See \code{\link{model.offset}}.}

\item{method}{a string with value \code{"qr"} for the column-pivoted QR
decomposition (exactly as base R) or \code{"chol"} for the LDLT Cholesky
decomposition (slower than LLT Cholesky but more stable)}

\item{tol}{threshold tolerance for convergence. Should be a positive real number.}

\item{maxit}{maximum number of IRLS iterations. Should be an integer.}
}
\value{
A list with the elements
\item{coefficients}{a vector of coefficients}
\item{se}{a vector of the standard errors of the coefficient estimates}
\item{rank}{a scalar denoting the computed rank of the model matrix}
\item{df.residual}{a scalar denoting the degrees of freedom in the model}
\item{residuals}{the vector of residuals}
\item{s}{a numeric scalar - the root mean square for residuals}
\item{fitted.values}{the vector of fitted values}
}
\description{
fast generalized linear model fitting
}
