% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{family.eglm}
\alias{family.eglm}
\title{family method for eglm fitted objects}
\usage{
\method{family}{eglm}(object, ...)
}
\arguments{
\item{object}{eglm fitted object}

\item{...}{not used}
}
\value{
returns the family of the fitted object
}
\description{
family method for eglm fitted objects
}
