% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{glance.eglm}
\alias{glance.eglm}
\title{Glance a(n) eglm object}
\usage{
glance.eglm(x, ...)
}
\arguments{
\item{x}{A \code{eglm} object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A one row data.frame providing model-level
summarizations (e.g. goodness of fit measures and related statistics)
}
\description{
Glance a(n) eglm object
}
\examples{

# load the mtcars dataset
data(mtcars)

# fit a linear model
mtcars_model <- eflm::eglm(am ~ wt, mtcars, family = "binomial")

# One row data.frame with model-level summarizatons
glance(mtcars_model)

}
