% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{glance.elm}
\alias{glance.elm}
\title{Glance a(n) elm object}
\usage{
glance.elm(x, ...)
}
\arguments{
\item{x}{A \code{elm} object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A one row data.frame providing model-level
summarizations (e.g. goodness of fit measures and related statistics)
}
\description{
Glance a(n) elm object
}
\examples{
# load the trees dataset
data(trees)

# fit a linear model on timber volume
trees_model <- eflm::elm(Volume ~ Girth + Height, data = trees)

# One row data.frame with model-level summarizatons
glance(trees_model)

}
