% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eflm-package.R, R/elm-wfit.R, R/elm.R
\name{linear_models}
\alias{linear_models}
\alias{elm.wfit}
\alias{elm}
\title{Efficient Fitting of Linear Models}
\usage{
elm.wfit(
  x,
  y,
  w,
  offset = NULL,
  method = "qr",
  tol = 1e-07,
  singular.ok = TRUE,
  reduce = TRUE,
  ...
)

elm(
  formula,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  reduce = TRUE,
  ...
)
}
\arguments{
\item{w, weights}{an optional vector of weights to be used in the fitting
process. Should be \code{NULL} or a numeric vector. If non-NULL, weighted
least squares is used with weights \code{weights} (that is, minimizing
\code{sum(w*e^2)}); otherwise ordinary least squares is used.}

\item{offset}{this can be used to specify an \emph{a priori} known component
to be included in the linear predictor during fitting. This should be
\code{NULL} or a numeric vector or matrix of extents matching those of the
response. One or more \code{\link{offset}} terms can be included in the
formula instead or as well, and if more than one are specified their sum is
used. See \code{\link{model.offset}}.}

\item{method}{currently, only \code{method = "qr"} is supported.}

\item{singular.ok}{logical; if FALSE a singular fit is an error.}

\item{\dots}{For eglm: arguments to be used to form the default control
argument if it is not supplied directly. For weights: further arguments
passed to or from other methods.}

\item{formula}{an object of class \code{"\link{formula}"} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of model specification are given under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object coercible
by \code{\link{as.data.frame}} to a data frame) containing the variables in
the model. If not found in \code{data}, the variables are taken from
\code{environment(formula)}, typically the environment from which \code{lm}
is called.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{\link{options}}, and is \code{\link{na.fail}} if that is unset. The
\sQuote{factory-fresh} default is \code{\link{na.omit}}. Another possible
value is \code{NULL}, no action. Value \code{\link{na.exclude}} can be
useful.}

\item{model, x, y, qr}{For \code{elm}: logicals. If TRUE the corresponding
components of the fit (the model frame, the model matrix, the response,
the QR decomposition) are returned. For \code{lm.wfit}: x is a design matrix
of dimension \code{n * p}, and y is a vector of observations of length n,
or a matrix with n rows.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of
\code{model.matrix.default}.}

\item{intercept}{logical. Should an intercept be included in the \emph{null}
model?}
}
\value{
an object of class "elm" that behaves the same way as the "lm" class,
see the function \link{lm}.
}
\description{
Efficient Linear Model (\code{"elm"}) is used to fit linear
models in an equivalent way to \code{"\link{lm}"} but in a reduced time
depending on the design matrix (see the DESCRIPTION).
}
\details{
Models for \code{elm} and \code{eglm} are specified symbolically.
A typical model has the form \code{response ~ terms} where \code{response}
is the (numeric) response vector and \code{terms} is a series of terms which
specifies a linear predictor for \code{response}. A terms specification of
the form \code{first + second} indicates all the terms in \code{first}
together with all the terms in \code{second} with duplicates removed. A
specification of the form \code{first:second} indicates the set of
terms obtained by taking the interactions of all terms in \code{first}
with all terms in \code{second}. The specification \code{first*second}
indicates the \emph{cross} of \code{first} and \code{second}. This is
the same as \code{first + second + first:second}, and exactly the same as
\code{"\link{lm}"} and \code{"\link{glm}"} from the \link{stats} package.
}
\examples{
elm(mpg ~ wt, data = mtcars)
}
