% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{logLik.eglm}
\alias{logLik.eglm}
\title{logLik method for eglm fitted objects}
\usage{
\method{logLik}{eglm}(object, ...)
}
\arguments{
\item{object}{eglm fitted object}

\item{...}{not used}
}
\value{
Returns an object of class \code{logLik}
}
\description{
logLik method for eglm fitted objects
}
