% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fglm}
\alias{predict.fglm}
\title{Predict Method for FGLM Fits}
\usage{
\method{predict}{fglm}(
  object,
  newdata,
  type = c("link", "response"),
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{a fitted object of class inheriting from "fglm"}

\item{...}{further arguments passed to or from other methods}
}
\description{
Obtains predictions and optionally estimates standard errors of
those predictions from a fitted generalized linear model object.
}
