% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{print.eglm}
\alias{print.eglm}
\title{print method for eglm objects}
\usage{
\method{print}{eglm}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{not used}
}
\description{
print method for eglm objects
}
