% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_methods.R
\name{residuals.eglm}
\alias{residuals.eglm}
\title{residuals method for eglm fitted objects}
\usage{
\method{residuals}{eglm}(
  object,
  type = c("deviance", "pearson", "working", "response", "partial"),
  ...
)
}
\arguments{
\item{object}{eglm fitted object}

\item{type}{type of residual to be returned}

\item{...}{not used}
}
\value{
a vector of residuals
}
\description{
residuals method for eglm fitted objects
}
