% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-integration.R
\name{sandwich_reexported}
\alias{sandwich_reexported}
\alias{estfun}
\alias{bread}
\title{Functions exported from \pkg{sandwich} to implement \pkg{fixest} methods}
\description{
Functions exported from \pkg{sandwich} to implement \pkg{fixest} methods
}
\keyword{internal}
