% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-blm.R
\name{summary.blm}
\alias{summary.blm}
\title{Summarizing Linear Model Fits}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{an object of class "blm", usually, a result of a call to blm}

\item{\dots}{further arguments passed to or from other methods}
}
\description{
This function is a method for class blm objects.
}
