% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-elm.R
\name{summary.elm}
\alias{summary.elm}
\title{Summarizing Linear Model Fits}
\usage{
\method{summary}{elm}(object, ...)
}
\arguments{
\item{object}{an object of class "elm", usually, a result of a call to elm}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
An object of class "list", see the function \link{summary.glm}.
}
\description{
This function is a method for class elm objects.
}
