% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fglm}
\alias{summary.fglm}
\title{Summarizing Generalized Linear Model Fits}
\usage{
\method{summary}{fglm}(object, ...)
}
\arguments{
\item{object}{an object of class "fglm", usually, a result of a call to fglm}

\item{\dots}{further arguments passed to or from other methods}
}
\description{
This function is a method for class fglm objects.
}
