% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.flm}
\alias{summary.flm}
\title{Summarizing Linear Model Fits}
\usage{
\method{summary}{flm}(object, ...)
}
\arguments{
\item{object}{an object of class "flm", usually, a result of a call to flm}

\item{\dots}{further arguments passed to or from other methods}
}
\description{
This function is a method for class flm objects.
}
