% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{tidy.eglm}
\alias{tidy.eglm}
\title{Tidy a(n) eglm object}
\usage{
tidy.eglm(x, conf.int = FALSE, conf.level = 0.95, exponentiate = FALSE)
}
\arguments{
\item{x}{A \code{eglm} object.}

\item{conf.int}{Logical indicating whether or not to include
a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence
interval if conf.int = TRUE. Must be strictly greater than 0
and less than 1. Defaults to 0.95, which corresponds to a
95 percent confidence interval.}

\item{exponentiate}{A logical indicating whether or not model terms should be presented on an exponential scale}
}
\value{
A data.frame summarizing component-level
information about the model
}
\description{
Tidy a(n) eglm object
}
\examples{
# load the mtcars dataset
data(mtcars)

# fit a generalized linear model
mtcars_model <- eflm::eglm(am ~ wt, mtcars, family = "binomial")

# summarize model coefficients in a tidy tibble!
tidy(mtcars_model)

}
