% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-broom.R
\name{tidy.elm}
\alias{tidy.elm}
\title{Tidy a(n) elm object}
\usage{
tidy.elm(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A \code{elm} object.}

\item{conf.int}{Logical indicating whether or not to include
a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence
interval if conf.int = TRUE. Must be strictly greater than 0
and less than 1. Defaults to 0.95, which corresponds to a
95 percent confidence interval.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A data.frame summarizing component-level
information about the model
}
\description{
Tidy a(n) elm object
}
\examples{
# load the trees dataset
data(trees)

# fit a linear model on timber volume
trees_model <- eflm::elm(Volume ~ Girth + Height, data = trees)

# summarize model coefficients in a tidy tibble!
tidy(trees_model)

}
